slightly cleaner install
